pyproject.toml
[project]
name = "agenda-cultural"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "playwright>=1.54.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "reflex>=0.8.6",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-coverage>=0.0",
    "ruff>=0.12.9",
]
